Automatic build of tagged commits with GitLab CI
To make a release, simply update the build.gradle file, tag the version
and push it. The CI will automically pick up the tag, build the JAR
and publish it as an artifact.

There's an issue in Gradle prior to 5.3 where non-ASCII characters in the
environment variable makes it crash. See:
https://github.com/gradle/gradle/issues/3117

To work around this without updating the used Gradle version, clear
the environment of any variable (mainly GitLab CI's) except PATH.